refactor: wizard types removed from spells lib
