+++ packages/wizards-of-react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@xstate-wizards/wizards-of-react",
-  "version": "0.5.23",
+  "version": "0.5.25",
   "author": "Mark Hansen",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
@@ -16,7 +16,7 @@
     "test:ci": "jest --runInBand --expand"
   },
   "dependencies": {
-    "@xstate-wizards/spells": "^0.5.25",
+    "@xstate-wizards/spells": "^0.5.26",
     "@xstate/react": "1.3.4",
     "date-fns": "^2.4.1",
     "lodash": "^4.17.20",
